// src/types.ts
export interface SourceFieldDescriptor {
  name: string;
  type?: string;
  description?: string;
  aliases?: string[];
  searchable?: boolean;
  aggregatable?: boolean;
  subfields: string[];
  nested_path?: string;
  object_array_path?: string;
  multi_field_parent?: string;
  preferred_exact_field?: string;
}
// ...
export interface SavedSecret {
  username: string;
  password: string;
}